Works with discord! But not how we really want it : (
Discord doesn' thave any input focused when the window isn't focused, and the bound key isn't being picked up when emulated to the window. I think this is due to how Discord is listening. It's not listening at window level but some other global level. Also it doesn't catch the key when trying to set the hotkey in the settings.